Move FunctionRegistry and FunctionCallException into the function package
