+++ src/main/java/io/burt/jmespath/function/FunctionCallException.java
@@ -1,4 +1,6 @@
-package io.burt.jmespath;
+package io.burt.jmespath.function;
+
+import io.burt.jmespath.JmesPathException;
 
 public class FunctionCallException extends JmesPathException {
   public FunctionCallException(String message) {
